requirements.txt
langchain
langchain-community
langchain-groq
chromadb
langchain-huggingface
PyMuPDF
python-dotenv